fix: reagruping code for clarity sake
