+++ android/app/src/main/java/org/ZingoLabs/Zingo/BackgroundSyncWorker.kt
@@ -1,18 +1,45 @@
 package org.ZingoLabs.Zingo
 
 import android.content.Context
+import android.os.Build
 import androidx.work.Worker
 import androidx.work.WorkerParameters
 import android.util.Log
+import androidx.annotation.RequiresApi
+import androidx.work.Constraints
+import androidx.work.ExistingPeriodicWorkPolicy
+import androidx.work.NetworkType
+import androidx.work.PeriodicWorkRequest
+import androidx.work.WorkManager
 import java.io.File
 import java.util.*
 import org.json.JSONObject
 import java.nio.charset.StandardCharsets
 import com.facebook.react.bridge.ReactApplicationContext
+import kotlinx.datetime.Clock
+import kotlinx.datetime.DateTimeUnit
+import kotlinx.datetime.Instant
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.atTime
+import kotlinx.datetime.toInstant
+import kotlinx.datetime.toLocalDateTime
+import kotlinx.datetime.until
+import kotlin.random.Random
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.days
+import kotlin.time.Duration.Companion.hours
+import kotlin.time.Duration.Companion.minutes
+import kotlin.time.DurationUnit
+import kotlin.time.toDuration
+import kotlin.time.toJavaDuration
 
 class BackgroundSyncWorker(context: Context, workerParams: WorkerParameters) : Worker(context, workerParams) {
 
+    @RequiresApi(Build.VERSION_CODES.O)
     override fun doWork(): Result {
+        // first scheduling the same task for tomorrow
+        BSCompanion.scheduleBackgroundTask()
+
         val reactContext = ReactApplicationContext(MainApplication.getAppContext())
         val rpcModule = RPCModule(reactContext)
 
@@ -40,7 +67,7 @@ class BackgroundSyncWorker(context: Context, workerParams: WorkerParameters) : W
             } else {
                 // this means the App is open,
                 // stop syncing first, just in case.
-                BSCompanion.stopSyncingProcess()
+                stopSyncingProcess()
             }
 
             // interrupt sync to false, just in case it is true.
@@ -92,6 +119,40 @@ class BackgroundSyncWorker(context: Context, workerParams: WorkerParameters) : W
         }
     }
 
+    private fun stopSyncingProcess() {
+        val resp = RustFFI.execute("syncstatus", "")
+        Log.i("SCHEDULED_TASK_RUN", "status response $resp")
+
+        var data: ByteArray = resp.toByteArray(StandardCharsets.UTF_8)
+        var jsonResp = JSONObject(String(data, StandardCharsets.UTF_8))
+        var inProgressStr: String = jsonResp.optString("in_progress")
+        var inProgress: Boolean = inProgressStr.toBoolean()
+
+        Log.i("SCHEDULED_TASK_RUN", "in progress value $inProgress")
+
+        while (inProgress) {
+            // interrupt
+            val resp2 = RustFFI.execute("interrupt_sync_after_batch", "true")
+            Log.i("SCHEDULED_TASK_RUN", "Interrupting sync: $resp2")
+
+            // blocking the thread for 0.5 seconds.
+            Thread.sleep(500)
+
+            val resp3 = RustFFI.execute("syncstatus", "")
+            Log.i("SCHEDULED_TASK_RUN", "status response $resp3")
+
+            data = resp.toByteArray(StandardCharsets.UTF_8)
+            jsonResp = JSONObject(String(data, StandardCharsets.UTF_8))
+            inProgressStr = jsonResp.optString("in_progress")
+            inProgress = inProgressStr.toBoolean()
+
+            Log.i("SCHEDULED_TASK_RUN", "in progress value $inProgress")
+        }
+
+        Log.i("SCHEDULED_TASK_RUN", "sync process STOPPED")
+
+    }
+
     private fun walletExists(): Boolean {
         // Check if a wallet already exists
         val file = File(MainApplication.getAppContext()?.filesDir, "wallet.dat")
@@ -107,43 +168,84 @@ class BackgroundSyncWorker(context: Context, workerParams: WorkerParameters) : W
 
 class BSCompanion {
     companion object {
-        fun stopSyncingProcess() {
-            val resp = RustFFI.execute("syncstatus", "")
-            Log.i("SCHEDULED_TASK_RUN", "status response $resp")
+        private const val taskID = "Zingo_Processing_Task_ID"
+        private val SYNC_PERIOD = 24.hours
+        private val SYNC_DAY_SHIFT = 1.days // Move to tomorrow
+        private val SYNC_START_TIME_HOURS = 3.hours // Start around 3 a.m. at night
+        private val SYNC_START_TIME_MINUTES = 60.minutes // Randomize with minutes until 4 a.m.
+        @RequiresApi(Build.VERSION_CODES.O)
+        fun scheduleBackgroundTask() {
+            val reactContext = ReactApplicationContext(MainApplication.getAppContext())
 
-            if (resp.lowercase().startsWith("error")) {
-                Log.i("SCHEDULED_TASK_RUN", "sync process STOPPED - no lightwalletd likely")
-                return
-            }
+            val constraints = Constraints.Builder()
+                .setRequiresStorageNotLow(false) // less restricted
+                .setRequiredNetworkType(NetworkType.UNMETERED)
+                .setRequiresCharging(true)
+                .build()
 
-            var data: ByteArray = resp.toByteArray(StandardCharsets.UTF_8)
-            var jsonResp = JSONObject(String(data, StandardCharsets.UTF_8))
-            var inProgressStr: String = jsonResp.optString("in_progress")
-            var inProgress: Boolean = inProgressStr.toBoolean()
+            // PRODUCTION - next day between 3:00 and 4:00 am.
+            val targetTimeDiff = calculateTargetTimeDifference()
 
-            Log.i("SCHEDULED_TASK_RUN", "in progress value $inProgress")
+            Log.i("SCHEDULING_TASK", "calculated target time DIFF $targetTimeDiff")
 
-            while (inProgress) {
-                // interrupt
-                val resp2 = RustFFI.execute("interrupt_sync_after_batch", "true")
-                Log.i("SCHEDULED_TASK_RUN", "Interrupting sync: $resp2")
+            // DEVELOPMENT - after 5 minutes.
+            //val timeFiveMinutes: Long = 5
 
-                // blocking the thread for 0.5 seconds.
-                Thread.sleep(500)
+            val workRequest = PeriodicWorkRequest.Builder(BackgroundSyncWorker::class.java, SYNC_PERIOD.toJavaDuration())
+                .setConstraints(constraints)
+                .setInitialDelay(targetTimeDiff.toJavaDuration())
+                //.setInitialDelay(timeFiveMinutes, TimeUnit.MINUTES)
+                .build()
 
-                val resp3 = RustFFI.execute("syncstatus", "")
-                Log.i("SCHEDULED_TASK_RUN", "status response $resp3")
-
-                data = resp.toByteArray(StandardCharsets.UTF_8)
-                jsonResp = JSONObject(String(data, StandardCharsets.UTF_8))
-                inProgressStr = jsonResp.optString("in_progress")
-                inProgress = inProgressStr.toBoolean()
-
-                Log.i("SCHEDULED_TASK_RUN", "in progress value $inProgress")
-            }
-
-            Log.i("SCHEDULED_TASK_RUN", "sync process STOPPED")
-
+            Log.i("SCHEDULING_TASK", "Enqueuing the background task - Background")
+            WorkManager.getInstance(reactContext)
+                .enqueueUniquePeriodicWork(
+                    taskID,
+                    ExistingPeriodicWorkPolicy.REPLACE,
+                    workRequest
+                )
         }
+
+        private fun calculateTargetTimeDifference(): Duration {
+            val currentTimeZone: TimeZone = TimeZone.currentSystemDefault()
+
+            val now: Instant = Clock.System.now()
+
+            val targetTime =
+                now
+                    .plus(SYNC_DAY_SHIFT)
+                    .toLocalDateTime(currentTimeZone)
+                    .date
+                    .atTime(
+                        hour = SYNC_START_TIME_HOURS.inWholeHours.toInt(),
+                        // Even though the WorkManager will trigger the work approximately at the set time, it's
+                        // better to randomize time in 3-4 a.m. This generates a number between 0 (inclusive) and 60
+                        // (exclusive)
+                        minute = Random.nextInt(0, SYNC_START_TIME_MINUTES.inWholeMinutes.toInt())
+                    )
+
+            val targetTimeTime = targetTime.time
+            val targetTimeDate = targetTime.date
+            Log.i("SCHEDULING_TASK", "calculated target time $targetTimeTime and date $targetTimeDate")
+
+            return now.until(
+                other = targetTime.toInstant(currentTimeZone),
+                unit = DateTimeUnit.MILLISECOND,
+                timeZone = currentTimeZone
+            ).toDuration(DurationUnit.MILLISECONDS)
+        }
+
+        fun cancelExecutingTask() {
+            val reactContext = ReactApplicationContext(MainApplication.getAppContext())
+
+            // run interrupt sync, just in case.
+            val resp = RustFFI.execute("interrupt_sync_after_batch", "true")
+            Log.i("SCHEDULED_TASK_RUN", "Interrupting sync: $resp")
+
+            Log.i("SCHEDULING_TASK", "Cancel background Task")
+            WorkManager.getInstance(reactContext)
+                .cancelUniqueWork(taskID)
+        }
+
     }
 }
